requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart
pyjwt
bcrypt==4.0.1
passlib[bcrypt]
sqlalchemy
sqlmodel
psycopg2-binary
python-dotenv
pydantic-settings